requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
python-docx
groq